Updates per Bill
